backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fillwise-backend"
version = "3.0.0"
description = "FillWise 3.0 — Legal Document Transformation Backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.29",
    "alembic>=1.13.1",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    # Validation
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.2.1",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Document processing
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    # HTTP client
    "httpx>=0.27.0",
    "ollama>=0.4.7",
    # Logging
    "structlog>=24.1.0",
    # YAML
    "PyYAML>=6.0.1",
    "jsonschema>=4.22.0",
    # Security
    "cryptography>=42.0.5",
    "slowapi>=0.1.9",
    # Metrics
    "prometheus-client>=0.20.0",
    # Utilities
    "python-dateutil>=2.9.0",
    "orjson>=3.10.1",
    "tenacity>=8.3.0",
    # ML / NLP (risk analyzer TF-IDF)
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.100.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning", "ignore::PendingDeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "ISC", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
